refactor: move ModalTemplate to shared dir
