--- src/screens/smu/components/modal/components/index.tsx
@@ -1,1 +1,0 @@
-export { default as ModalTemplate } from './ModalTemplate';
